scripts/requirements.txt
psycopg2-binary==2.9.9
h3==3.7.7
numpy==1.26.4
requests==2.31.0
